requirements.txt
cycler==0.10.0
decorator==4.4.1
dj-database-url==0.5.0
Django==2.2.7
django-heroku==0.3.1
django-storages==1.8
djangorestframework==3.10.3
gunicorn==20.0.2
imageio==2.6.1
imutils==0.5.3
kiwisolver==1.1.0
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==6.2.1
psycopg2==2.8.4
pybase64==1.0.1
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
scikit-image==0.16.2
scipy==1.3.3
six==1.13.0
sqlparse==0.3.0
whitenoise==4.1.4
